Bug 704212: Fix race condition in fz_advance_glyph.
Suppose one thread calls fz_advance_glyph, determines that we should
be using an advance cache, but one doesn't exist. It will then
allocate it, and start to fill it.

If at that point another thread calls fz_advance_glyph, it'll see
the newly created advance cache, and access it regardless of the
fact that it's newly only partially filled.

The way to avoid this is to hold the FREETYPE lock around the
use of the advance_cache.