Fix bug in hex_from_color.
If no color is specified (when n is 0) then the color is either
transparent (for annotation borders) or default to black (for
variable text).

In the UI we convert the color array to an RGB hex in order to
display the color name. Use transparent/none rather than black when the
color is not set.