Fix fz_copy_pixmap_area_converting_seps to cope with spot cases.
When copying from CMYK + Spots to CMYK, we were incorrectly falling
into the 'just copy it' clause, resulting in spots being dropped.

This meant that:

 mutool draw -o spot1.psd -O 1 REWE2.pdf

(using REWE2.pdf from Bug 701031) was coming out in greyscale.

With this change in both the above command and:

 mutool draw -o spot2.psd -O 2 REWE2.pdf

appear almost identically in Photoshop (and are visually identical
to what Acrobat renders).