Fix some premultiplied problems.
First, fix a typo in lcms2 that caused a non-premultiplied
transform to be used when a premultiplied one was required.

Next, allow for the limitations in lcms2's premultiplication
support by reinstating the 'do it by steam' routines that
undo/redo premultiplication for some cases.

LCMS2 can handle premultiplication in the cases where the
number of spot channels does not change. In rare cases (such
as folding spots down to process colors after overprint simulation)
we can need to operate not in this way.

We now make use of the faster lcms2 processing where possible,
and do it ourselves otherwise.