Bug 703138: js: Add interfaces for getting/setting field types.
Previously retrieving the type of a field was not supported.
PDF objects 848 and 988 in PDF in bug 703138 has additional
javascript actions that call getField() and then determines its
type through the type property. Instead of the type evaluating
to e.g. "checkbox", the type would always evaluate to undefined.
This caused the javascript code to break and not limit the
checkboxes to be mutually exclusive.