Bug 704341: Fix redaction of image masks.
Previously image masks were not handled specially when being
redacted. Because image masks a stored inverted in the PDF, when
redacting them we need to invert the samples, perform the redaction
including erasing pixels and drawing black boxes as requested. Finally
we need to remember the image mask flag so that the image is written
correctly when the PDF is saved.