Bug 704485: Rejig linearization process during clean.
The file supplied in the bug has a large Outlines structure.
The existing linearisation code recurses through the structure
using an fz_try at each stage, so we overflow the exception stack.

The fix for this is to revise the code so as not to use an fz_try
at each stage. This means we need some way to 'unpick' the marking
of objects if we throw an exception. We do this by using a new
'mark_list' structure that keeps a list of the objects that need to
be unmarked in the event of a throw.

This solves the problem, but leaves us with a process that takes
a long time to run (5 minutes for the 'small' example on the bug).
This is due to the recursive traversal of all the file objects
traversing the same objects several times (if it reaches them by
different routes).

We fix this by adding a condition that allows us to bale out of
recursing onwards from marking an object if that object has
already been marked with the given flags (as presumably all its
subobjects would have been marked with those flags in any previous
visit).

As part of this fix, we add a new pdf_mark_list structure, and
function to init/free/push/pop stuff onto this list as we mark/unmark.