PDF iterator fix: next on 'empty' iterator was wrong.
If we tried to moved 'next' on an iterator formed on an empty
set of outlines, we got outselves into the situation where
we had MOD_AFTER and iter->current == NULL, which should never
be allowed.

Fix 'next' to spot the attempt and return the appropriate
return code.

We were seeing this with:

  mutool merge -o out.pdf pdf_reference17.pdf 52

Same fix for the other movements.