Bug 704337: Fix memory overread while cleaning a fuzzed file.
When cleaning the supplied file and attempting to remove duplicate
objects, it's possible to be tricked into reading beyond the end
of the use_list. We fix this by ensuring that use_list should always
be long enough.

A quick review of the code suggests a few other places where we
could do with similarly extending the lists, so fix those too.

Thanks to Cenzhang for the fuzzing, and an initial proposed
fix.