scripts/mupdfwrap.py: fixed bug in generation of fz_link iterators.
We were creating a LinkIterator by passing a raw fz_link* to LinkIterator's
'Link m_item' member. Link's constructor follows our convention where raw fz_*
pointers passed to constructors are assumed to be owned, so does not call
fz_keep_link(), so we were not incrementing the reference count as we should
have done.

The fix is to for Link::begin() to pass *this to LinkIterator's constructor
instead of this->m_internal.