Fix down conversion from CMYK+spots+alpha to RGB+alpha.
This case is hit by:

 mutool draw -o out.png -c rgba map.pdf

(where map.pdf is a private PDF file that I cannot share). It's a
file with a spot color, where the top level group is RGB.

Because of the spot color (presumably) we render internally to
cmyk+spot+alpha.

We then finally convert this down to rgb+alpha in fz_draw_close_device
by calling fz_copy_pixmap_area_converting_seps().

This filters down, and ends up in the last section of
fz_convert_pixmap_samples, which calls fz_icc_transform_pixmap.

The logic in fz_convert_pixmap_samples didn't quite match
the logic in fz_icc_transform_pixmap about whether premultiplication
was required or not. Accordingly the conversion was happening with
each thinking the other was responsible for handling the
premultiplication.

Ensure the expectations match up.