Fix obj streams with no whitespace.
A customer has supplied us with a file that uses compressed object
streams, where a series of objects are all 'null'.

The stream for this is of the form:

 ...nullnullnullnull...

The current code attempts to parse and doesn't recognise the end of
the area of the stream for this object, so finds a keyword called
"nullnullnullnull..." which upsets our parsing.

The fix, implemented here, is to use a range filter to restrict the
parsing.