Add fz_link_dest type.
This is a new API for conveniently handling more detailed link actions.

The type is based on the PDF actions and the "Adobe Acrobat SDK Parameters for
Opening PDF Files" specification.

We add functions to parse and format internal link URIs for PDF document types
embedding all this information.

There is a new function fz_resolve_link_dest that can be used to resolve links
to a fz_link_dest, which supplants the old fz_resolve_link.

For editable document formats (PDF only so far) there is also a document
specific function to format a link URI called fz_format_link_uri that takes a
fz_link_dest.

For most clients, the URI should be treated as an opaque identifier and be
parsed and created using the mentioned functions.

The Java library and mutool run environment also has new corresponding classes
and methods.