Fix assert when reading an SVG with a 0 opacity group.
When picking the painter routine to use for group contents, if
dst->alpha is 0, we return NULL. Unfortunately, we assert that
the routine is non-NULL despite actually otherwise handling it
correctly. The fix, implemented here, is to update the assert
condition.

This has never been tripped before, because who'd be dumb
enough to output an SVG file with an object with 0 alpha in it?
Evidently, us; this SVG file was generated from our svg-writer,
so update that to avoid writing 0 opacity objects in most cases.
We still have 1 case (begin_group) where it's not trivial to
elide the entire group.