Create /Outlines as an object, not a document catalog internal dictionary.
According to the specification /Outlines in the document catalog
must be an indirect reference to an outline dictionary. Previously
MuPDF would store the outline directionary directly as an internal
dictionary in the document catalog. Moreover MuPDF would store a
copy of the parent dictionary under /Parent in each outline item node.

After this commit the document catalog /Outline entry is an indirect
reference to an outline dictionary and each outline item node's
/Parent entry uses indirect reference to its parent node.