Bug 704778: Fix spot color downconversion.
The file has an output intent, and so the last step in rendering
the file involves mapping from RGB+Spot to just RGB. The code for
doing this was inverting the spot value.

Thus a 0 value for the spot was being mapped to a full value.

I have done more tests with the cluster, and I've now tested 3
out of the 4 routes through this code. All now behave as we
expect. The 4th one looks plausible by desk checking.

Possibly we can now common up 2 of the 4 cases, but I'll leave
this for later.