Bug 704786: Support repairing prev pointers in outline tree.
The reported file has a broken outline.

It contains an outline node (777 0 R) that has children. Its /First
entry points to a child node (778 0 R) that has a /Prev entry. This
is unexpected according to the specification.

Moreover that /Prev entry points to an outline node (776 0 R)
that does not have a /Next entry pointing back to the child node
(778 0 R), but rather is part of a different linked list of outline
nodes.

Let MuPDF handle this by assuming that all /First and /Next entries
are correct and deduce what /Prev ought to point to. If no /Prev
entry is expected then remove it if it exists. If the /Prev entry
points to the wrong node then point it to the expected outline node.