Support fixing last pointers in outline tree.
MuPDF assumes that all /First and /Next entries in outline nodes
are correct. By following the linked list of /Next pointers until
the end it is possible to deduce what the parent's /Last entry
ought to point to.

This commit adds code to warn if the parent's /Last entry is
unexpected and also fix its /Last entry to point to the expected
outline node.