Bug 704289 (Continued): Fix inability to fill in fields in repaired docs.
This is only passingly related to the original bug, but was spun out
of it due to finding the issue while debugging that bug.

If you edit the file so that it requires a repair on loading, you'll
find that the value of fields in the file can be edited/changed, but
their appearance is never updated.

This is because 'prepare_object_for_alteration' refuses to do anything
if 'doc->repair_attempted'. We need to prevent
prepare_object_for_alteration doing anything while we are actively
repairing a file, but once this is complete, we should allow it to
behave as before.

Accordingly, we add a new doc->repair_in_progress field, and work
on that, rather than on repair_attempted.