Mudraw: fix text extraction when used at dpi's != 72.
The page bbox used to clip extracted text was always unscaled.
Hence extracting a page at 200dpi would clip to only a quarter
of the page being extracted. This is a greater problem with
OCR, where higher dpi is required!

In a similar vein, make mudraw honor the resolution/rotation for
other extraction types too.