Bug 703988: Fix H/V word direction detection in OCR.
Not actually to do with the bug itself, but noticed while debugging it.

If the first paragraph in the test file is redacted then the "The"
in the following text ("The Marketplace Manager is...") is detected
as vertical movement. This is due to typos (and improper ordering)
within queue_word.