Extract CFF table from OpenType fonts in PDF.
FreeType treats glyph indexes in "pure" and "embedded" CFF CID-fonts
differently. For embedded CID-fonts it treats GIDs as a raw index into
the CharStrings array, but for pure CID-fonts it maps the GID via the
charset table.

PDF requires CID-fonts embedded in an OpenType to map the GID
via the charset same as for pure CID-fonts. See page 438 in pdfref17.

There is currently no way to ask FreeType for this behavior.
Instead we extract the CFF subtable and pass it as a pure CFF font
to FreeType, to get the desired GID/CID mapping behavior. PDF does not
need or use any of the other OpenType tables in the font, so this is
safe.