Throw exception if Size entry in trailer is indirect.
Previously when MuPDF encountered an indirect /Size entry while
parsing the xref it would try to resolve the indirect object before
having fully read the xref causing a crash.

Now MuPDF ensures that the /Size entry is not an indirect object
as this is not allowed according to the PDF specification.

Fixes OSS-fuzz issue 46372.