Only store font descriptor for type 3 fonts once.
When MuPDF is interpreting page contents into a display list and
the contents refer to a type 3 font it loads the font. MuPDF
first queries the store to check if the font descriptor has been
cached. If so it is returned and used during rendering, otherwise
the font descriptor is loaded from the the document file.

When loading the font descriptor, the CharProcs stream for each
glyph in the font is read and interpreted into a display list.
This happens when pdf_load_font() calls pdf_load_type3_glyphs()
which in turn calls fz_prepare_t3_glyph() for each glyph.

Interpretation of the font's CharProcs streams into display lists
is done so that subsequent processing of the page display can
solely rely on the store for font information and avoid having to
re-parse the document file. So caching the font descriptor is
necessary. Once the font descriptor has been loaded, it is
unconditionally inserted into the store.

Type 3 fonts may have cycles which cause problems:

 * a CharProcs stream for one glyph recursively referring to
   its own glyph.

 * and a CharProcs stream for one glyph referring to that of
   another glyph in the same font.

The first type causes obvious cylic behaviour when interpreting
the CharProcs stream of the glyph, the second type does not.

MuPDF already handles the first type of cycle by doing two things
in fz_prepare_t3_glyph() and fz_render_t3_glyph_direct():

 1) checking whether the CharProcs stream is NULL before trying
    to process the glyph,

 2) and temporarily setting a glyph's CharProcs stream to NULL
    while its stream is being processed.

This breaks cyclic references since the CharProcs stream will be
NULL at the beginning of the nested call, causing the nested
reference not to be rendered. This breaks the cycle for an
individual glyph but is not enough to prevent the type 3 font
from being loaded recursively.

Both types of cycles cause asserts in debug builds due to repeated
caching of font descriptors in the store.

The intial loading of the the type 3 font causes recursive loading of
the type 3 font due a CharProcs stream refering to the font itself.
The font descriptor is first inserted into the store in the nested
call. Once all glyphs have been processed in the top-level call it
will again insert the font descriptor into the store.

In release builds asserts are not enabled, but they are in debug
builds. When inserting objects into the store it asserts that there
must not already be an existing entry. So when type 3 fonts with
cycles of the second type cause the font descriptor to be inserted a
second time this assert is triggered because it already exists.

MuPDF now removes the font descriptor potentially cached in the nested
call. This means that later on when pdf_load_font() tries to insert
the font descriptor again there will not be any existing entry and the
assert is not triggered.

This fixes OSS-fuzz issue 38139.