Reset radio button to a valid default value, otherwise Off.
Previously when resetting a radio button field group to its default
value all dependent radio button fields would inherit this and set
their value to the default value.

Consider a radio button field group with default value /accepted and
two dependent radio button fields one with the appearance states /Off
and /accepted and the other field with the appearance states /Off and
/declined. When the field group was reset its value changed to
/accepted. This then propagated to both of the dependent fields, but
only one of them has the appearance state /accepted. So setting the
value in both radio button field to /accepted is wrong.

With this commit, a radio button field can only be reset to values
that exist in its normal appearance state dictionary. In the example
above it means that the radio button that lacks the normal appearance
state /accepted will be set to /Off instead.