Make sure that excluded fields are excluded when resetting form.
pdf_reset_form() takes a list of fields and an exlusion flag. If the
exclusion flag is cleared the fields in the list should be reset. If
the exclusion flag is set all fields EXCLUDING those in the list
should be reset.

This should be accomplished by calling specified_fields() which should
return an array of fields to reset. specified_fields() receives the
exclusion flag and the list of fields and will add /Exclude null
entries to field dictionaries that should be ignored.

Later on specified_fields() calls add_field_hierarchy_to_array() to
filter out the fields that should be reset into a resulting array.
This detection consists of looking for the /Exclude null entry in each
field's dictionary. The problem is that since the value of /Exclude is
null then it cannot differentiate between no entry and an /Exclude
null entry.

This is likely a regression from when PDF_NULL became == NULL in
commit 706ed587b875fab178fbfc2b60e1a49527e89b80.

The approach with the /Exclude flag has the drawback that all
PDF field object will have been changed regardless of whether it
has been reset or not.

This commit changes the approach to for each field scan through the
(likely short) list of fields to include/exclude. This avoids
misinterpreting the /Exclude flag and avoids unnecessarily dirtying
PDF field objects.