HTML layout: Add "-mupdf-leading" css property.
Add -mupdf-leading to the properties in css-properties.gperf and rebuild
css-properties.h from that.

Update the CSS property routines to cope with 'undefined' properties.
The -mupdf-leading property is initially undefined, and only set when
met within the file.

Update the layout routines to honour -mupdf-leading. If -mupdf-leading
is set, then we always 'advance' lines by '-mupdf-leading', rather than
the maximum of the lineheight, and the text height. We still position
the text according to the lineheight.