Split HTML story parsing into 2 phases.
Previously we used to parse HTML text to xml, and thence to boxes
all in the fz_new_html_story() call.

Here we move to only parsing HTML text to xml in the
fz_new_html_story() call, and we go from xml to boxes on the first
layout.

This opens us up to being able to do fz_new_html_story() on an
empty buffer, and then programmatically changing the XML in a DOM-like
way before converting to boxes.