Tweak makefile for Tesseract flags.
Firstly, make sure that we supply -std=C++11 as a flag when
building tessocr.cpp. gcc on MacOS was defaulting to an older
version. We do this using a TESSERACT_LANGFLAGS variable so
that this can be overridden from the command line if required
(say, because someone is using something other than gcc).

Secondly, ensure that we pass the same set of predefines to
both tessocr.cpp and the tesseract build itself, so that
tesseract headers always see a consistent set of defines.