Add XML dom functions.
In here, we rework the fz_xml structure a bit. We remove the
fz_xml_doc type in favour of a new union within the fz_xml
structure. We typedef the two to be equivalent for backward
compatibility.

The topmost structure in any fz_xml tree has a NULL up pointer
(reasonably enough). We use the nullness of this pointer to
indicate that it is a 'doc' structure, rather than a 'node'
structure. 'doc' structures simply hold a down pointer (that
points to the root 'node' in the tree), together with the pool
pointer used for all allocations in the tree, and a reference
count for the tree as a whole.

Behaviour of the 'node' structures within the tree remain
broadly unchanged, apart from the fact that nodes that have
been created, but not linked into the tree still have an 'up'
pointer that points to the 'doc' element, despite not appearing
in any of the doc's child nodes.

Aside from changing the APIs to not use the fz_xml_doc type
any more, we remove an unnecessary doc parameter from
fz_detach_xml.

Finally, (and the ultimate intent of this commit!) we add a
set of functions to implement DOM-like access for HTML within
XML.