Bug 705368: Fix HTML layout of text around images.
In the commit to add '-mupdf-leading', I'd altered measure_line
so that it returned both the height of the box and the vertical
advance.

In the absence of '-mupdf-leading', I'd forgotten to update
vertical advance in the same way that height was updated when
encountering a taller node (such as an image).

We no longer need to test for whether the value of leading
is defined or not, because regardless node->h will have been
set to be either leading, or from the em value prior to this
being called.

In the previous version of this code, we'd taken the last
value of leading that we met on a line as being the leading
for the complete line. We alter that to be that we take the
largest one that we meet.