mutool run: Remove superfluous (and confusing) calls to close device.
There are basically 2 cases in the javascript handling of devices;
native devices and javascript implemented devices.

If we are using native devices the code does not call _close() when
running content to the page. The caller is left to do that themselves
by calling device.close() (which ends up in ffi_Device_close which
calls fz_device_close). All is good with the world.

If we are using javascript devices, then each 'run' routine wraps the
javascript device in a magic object so that C calls to the members
get converted to js ones. In this case the 'run' routines currently
call fz_close_device at the end. This is somewhat confusing, because
a) it's not what would happen with native devices, and b) the
implementation of fz_close_device doesn't actually call through to
any javascript 'close' method.

For clarity, we just remove those here.