Bug 705438: At end of unpack_arguments() call, pass NULL instead of 0.
A zero will through integer promotion be of type int, but what
unpack_arguments() actually expects is a pointer. As along as the
sizes of int and pointers are the same this causes no issue. But
on 64-bit platforms where pointers are 64 bits wide and int is
32 bits wide, the argument list will not be initialized fully.
In these situations the unpack_arguments() function will try to read
64 bits despite the trailing 0 only having initialized 32 bits at
the end of the argument list, possibly causing a crash.

Passing NULL will, since it is a void* pointer, will cause all
64 bits to be fully initialized fixing the problem.