svgdevice: Nest defs when necessary.
Sometimes we need to define something while in the middle of another
definition. Most of the times, these things use "safe" definition tags such as
clipPath or mask, but if it's a complex item that needs a <g> tag such as when
we define our pattern tiles or images, we should protect the nested definition
with a <defs> tag.