svgdevice: Don't put images in defs.
We don't need to put an image into <defs> in order to reuse it. When reusing
images, emit the first image instance directly, but remember the ID so we can
reuse it later.

This is safe because the <image> tags we emit don't contain any instance
specific attributes. The transform is always applied with a wrapping <g> tag.