Bug 705635: Add forward page map to improve merge performance.
Previously, we'd have a reverse page map (to map from object to
page number), but not a forward one (from page number to object).

This meant that looking up the destination of every outline object
in a document in turn (which are almost always in monotonically
increasing order) requires us to slave through a larger and larger
prefix of the page tree each time.

We therefore introduce a forward page map to match the reverse
page map.

This alone doesn't solve the problem in the bug though, as we
create the page map, and then destroy it before it can be made
use of, for fear that it'll be left caching outdated values.

Accordingly, we update 'pdf_load_page_map' and 'pdf_drop_page_map'
to increment/decrement a nesting value, and the map is only
dropped when the nesting value returns to zero.

It is the caller's responsibility to ensure that nothing changes
the underlying page tree between those 2 calls.