Return error, not success when unable to lock native device resource.
Commit 3bd8ce45a9a29c8131deab37df17c24b1ece162b accidentally changed
the return value upon thrown exceptions in the locking function.

This meant that even if the AndroidDrawDevice failed to lock the bitmap
it would set the pixmap sample pointer to NULL, throw an exception
and then return success to the caller. The caller newNativeAndroidDrawDevice()
detects errors based on the return value, because it indicates success
it would proceed to clear the pixmap samples. Since the samples pointer
was set to NULL this would generate a segmentation fault.

By instead indicating failure in the return value the caller will stop
using the pixmap samples, free all resources and propagate the error.
This is implemented in this commit and thus avoids the segmentation fault.