Ensure AndroidDrawDevice is destroyed, even upon exception.
Without this the AndroidDrawDevice and any resources on its draw stack
will remain until the Android JVM decides to garbage collect the
unreferenced AndroidDrawDevice object. This may take a long time, hence
it is prudent to call AndroidDrawDevice.destroy() when one knows that
the device object and its resources will not be used again.