Bug 705691: Fix SEGV in fz_drop_html_story().
Credit to Julian Smith for doing the legwork in finding this one.

As part of the fz_html_story() work, I refactored the HTML reading
so that both fz_html and fz_html_story types are 'derived' from
an fz_html_tree type. Unfortunately, I failed to note this well
in the code, which meant it wasn't immediately obvious.

It was so "not-obvious", in fact, that I then rejigged structure
contents so that fz_html_tree wasn't the first part of the derived
structure any more. This meant that BadThingsHappened(TM) when
the structure was freed.

Here, we move over to the standard way that derived classes are
done within MuPDF in the hopes that it'll catch out fewer people
(including myself!) in future.