Only generate appearance for standard stamp annotations.
Handle dirty tracking of Stamp annotations where the rectangle is
changed but the AP is not regenerated.

Deviate from the spec to allow custom Stamp annotation appearances.
Only regenerate if the Name is explicitly set to one of the standard
names. Don't treat a missing property as the default "Draft".