HTML Layout: Fix top spacing problem with restartable layout.
Firstly, we weren't using advance_for_spacing for the top
spacing, meaning that if the top spacing for a box would
push us onto a new page, we'd end up starting part way down
that page, rather than at the top.

With restartable layout, we weren't spotting that we'd
pushed to the new page, and hence weren't stopping the
layout. This meant we continued to layout until we hit
the next page, meaning that content was lost. Fix this
by consulting 'eop' returned by advance_for_spacing.