HTML Layout: Correct restart point if no nested content fitted.
When doing restarting layout, if BOX1's first child is BOX2, and no
content from BOX2 can be fitted, then we want the restart point
to be BOX1, not BOX2.

Otherwise, when we enumerate the positions for a page BOX1 will
appear to be on the previous page to BOX2.

We were doing this for boxes with borders, when we should be doing
it for all boxes. We were probably only doing it for boxes with
borders, because those are the only ones I could visibly see it
being required for.