Restartable HTML: Fix background color when restarting in a box.
If we have a box A with a background color, whose content is
sufficiently large to make it run over multiple pages, we are
currently only drawing the background color on the first page.

This is because, when we traverse the tree on the second or
subsequent pages, looking for the 'end-of-skip' position, this
will be within A. So we will still be skipping at the point
at which we examine A for drawing.

Accordingly, we change the logic for the background box
rendering to no longer rely on skipping, and instead to look
at whether the box is zero height (all fully skipped boxes
should be zero height).