Optimise (and correct) "overflow-wrap" handling.
Tor points out that this can be done without a dedicated
FLOW_WORD_WRAPPED node type, meaning we can use reuse
FLOW_WORD.

Also, tweak the logic so that:

   <p style=\"overflow-wrap:break-word\">This WordIs<span style=\"color:red\">ExtremelyImprobablyStupidlyMonumentallyActually</span>QuiteLong.\n"

is correctly recognised as being a single word.

Also, fix a problem spotted by Tor: in walk_string, in the
'quickshape' case, I was thinking that harfbuzz wasn't involved,
and so I was setting the cluster value myself. Turns out that
harfbuzz is still being called, so I don't need to set this at
all.