Bug 705838: Improve resilience to compression bombs.
When using fz_read_best to read from a stream, now allow it to take
a 'worst_case' value. If this is set, then don't trigger the
compression bomb code until we pass this value.

Accordingly, feed that value in from the image reading code with the
full size of the image, and leave it set to 0 elsewhere.