Remove needless uses of fz_cleanname.
If we call fz_cleanname and then just pass the path into
an archive function, this is now pointless, as the archive
functions all cleanname for themselves.

We can therefore remove 4 instances of this.

Other instances are still kept, because the cleaned paths are
used elsewhere instead of/as well as being passed to archive
functions.