Bug 705980: Use renumbered object when refering to cleaned Encrypt dict.
When keeping encryption while cleaning a PDF file MuPDF expected that
the encryption dictionary stays at the same object number even in the
cleaned file. This is true as along as the xref is not compacted.
Xref compaction will renumber objects, so when writing out the new
trailer refer to the Encryption dictionary object by the new object
number.