Fix color-filtering with separations.
Previously, if we were told to go into a separation space:

 /CS0 cs

and we rewrote that (say) to be RGB, when we then came to do:

 0 scn

we'd be trying to set the 'amount' of RGB, rather than the amount
of the original separation space.

Accordingly, we now store the untransformed colorspace in the
filter, so we can work from the correct source each time.