Bug 706094: Fix mutool clean garbage collection.
A while ago, we spotted that pdf_objcmp was comparing any 2 stream
objects as being identical if the dict portions matched, without
considering any differences in the stream sections.

We didn't want to take the hit of having pdf_objcmp compare stream
contents, so we modified the code so that any stream objects would
always differ.

Unfortunately, mutool clean was relying on stream dicts being matched
so it could do its own comparison of the stream contents. This meant
that mutool clean with garbage >= 4 was failing to properly common up
repeated streams.

To solve that here, we have introduced a new pdf_objcmp_deep that
performs as pdf_objcmp does, but also checks the stream contents.
We make mutool clean (actually removeduplicateobjs in pdf-write.c)
call that, and correct operation is restored.