html: Copy "href" attribute to child boxes directly.
1) We no longer need to recursively look upwards through the box
tree to find the href for boxes and flow nodes.

2) This works around our funky nesting of block/flow boxes.

Since we only hold on to the inline boxes for styling purposes inside
of flow nodes, the boxes don't nest in the same way as the elements.
This can cause funny effects if an inline element contains a block
element, such as in the following example:

<a href="http://example.com/">
<h1>Block element</h1>
</a>