Pass 'model' colorspace into fz_render_stroked_glyph.
This avoids various warning messages and incorrect renderings
with type3 glyphs.

This was seen with ContentStreamCycleType3insideType3.pdf
where a Type3 font is used with 'Stroke and Fill' mode. For
small sizes, the fill caused it to be rendered once and cached.
Over 220 dpi, it won't fit in the cache so the stroke tries to
rerender it, and is confused by model being NULL.