Miscellaneous font embedding tweaks.
Every fz_font gains 2 new flags:

'never_embed' is set from the flags in the font to avoid us
ever embedding a font that does not give permission.

'embed' is set according to the source of a font. By default
base14 fonts and inbuilt fonts have this set to false. Fonts
loaded other than this have it enabled.

Users can override the 'embed' flag for a font using a new
function, fz_set_font_embedding.

We also keep track of which subfont of a ttc is in use in a
new fz_font 'subfont' field.

If we are asked to embed a font that happens to be one from
a ttc, we now extract the ttf from the ttc so we can embed
that correctly.