Fix pdf_objcmp's stream detection.
pdf_objcmp now checks whether its inputs are streams or not
as part of its operation. In order to know whether an
object is a stream, we have to look it up in the xref.

This lookup can trigger a 'solidification' of the xref,
which requires an allocation. In such cases, we know we're not
a stream anyway, so it's largely pointless.

pdf_objcmp can be called in situations where the alloc lock
is taken already (in particular by pdf_cmp_key), so this is
problematic.

Accordingly, we introduce a variant of pdf_get_xref_entry
(pdf_get_xref_entry_no_change) that never solidifies the
xref - merely returning NULL in cases where it would be
required. This is sufficient for our purposes.

We update pdf_objcmp (or rather do_objcmp) to call that instead.