Bug 706381: Tweak xref allocation to reduce pathological reallocs.
Firstly, we update pdf_xref_find_subsection to distinguish between
4 different cases, rather than 3.

Previously we had coped with:

 a) the required range of objects being in a single xref subsection,
 b) the required range of objects not being in any xref subsection.
 c) anything else.

In this code, c) would have resulted in reallocing, copying, and
freeing. This would have been triggered repeatedly by the example
files attempt to add a single new object onto the end of the xrefs
repeatedly.

Now we cope with:

 a) the required range of objects being in a single xref subsection,
 b) the required range of objects not being in (or subsequent to)
    any xref subsection.
 c) the required range of objects 'extending' an xref subsection.
 d) anything else.

This enables us to spot the case triggered by the example file, and
to efficiently grow the subsection. It results in a nicer, more
efficient structure in memory in that we get a single solid xref
subsection of size n, rather than n subsections of size 1 in a list.