Bug 706412: Fix performance regression with loading PDF outlines.
In 1.19.0, we used to load outlines from PDF files in one hit.
As part of loading URIs in outline entries, we'd need to map
from page object to page number. To ensure this was speedy we'd
cache the lookup in a 'reverse page map'. We'd need to drop this
map once we'd loaded the outline, to allow for people modifying
the page map.

Shortly after 1.19.0, we introduced an outline iterator mechanism.
This meant that we no longer (necessarily) load the PDF outlines
in a single hit.

When an outline iterator is created, we would load the reverse
page map, and sanity check the outlines for errors. Before we return
the created iterator to the caller, we'd have to drop the reverse
page map lest modifications were done to the document.

So, even if the standard system function was then used to load the
outlines from the iterator (as is the case with PyMuPDF), this would
happen with us having thrown away the cache, and all the lookups
would be slow. This is what is happening in the bug.

So, the fix implemented here is to leave the reverse page map in
place when we exit, and to discard it whenever we modify the
document.

BUT... there is a wrinkle with this. As we parse objects from the
file, we build them in memory using the same functions as a programmer
would use to modify the doc, so we can actually destroy the reverse
page map just by reading objects from the file (such as happens
while we, for example, read the reverse page map).

Accordingly, we add a doc->caching_object flag which means
"I am currently caching an object from the file". If this is set
then we don't drop the reverse page tree.