Bug 706412 followup: Further improvements to rev_page_map discarding.
Firstly, we can lose the 'doc->caching_object' flag, because in all
the circumstances in which this would be set, we know that the
parent object number will be 0.

i.e. while parsing an object from the document, and constructing it
the document will not yet have been linked into the xref tree.

This also captures the lexing of objects within pdf_process_stream
that I'd missed before.

Continuing in the same vein, we can avoid dropping the reverse
page map when we are making changes within local_xrefs (so
annotation synthesis can no longer trigger it).

Repairs made to the document as a result of testing the outline
structure (which happens far more than we'd like) will still
trigger dropping rev_page_map, so we introduce a new
'non_structural_change' document flag that we can set while making
such repairs. We no longer drop the page_ref_map if this is
set.

All this is still slightly less than ideal in that we will
unquestionably drop the rev_page_map in cases when we don't need
to. It is safe, but not optimally fast. We can continue to think
about better solutions.