Bug 706123: Cope with EXIF orientation info in SVG files.
PDF ignores any EXIF data, and so does not honour the orientation
setting. Unfortunately, SVG does the opposite.

This means that when we convert a PDF to an SVG, any EXIF data in
the PDF which would previously have had its orientation data
ignored, suddenly finds it being honoured. As such, we now sanitize
the JPEG data as we output it. This sanitization process may be
extended to do other things too in the future (in particular to
tag for whether we should invert cmyk jpegs or not).

Also, when we read an SVG, we now look for orientation data and
honour it. We were already doing this in the image document handler.

Thus the convention remains that when JPEG data is passed over the
device boundary, any EXIF data is ignored.