Bug 706426: Workaround strange text order in extraction.
The example file has a load of text that purports to be Arabic.
Unfortunately, while it appears as arabic, the ToUnicode tables
for many (but not all) of the fonts in the document are horribly
broken, and so when cut and pasted, you get garbage characters.

Stranger still, in the lines where such text is used, the
characters within words are sent left to right, but the words
(and the spaces between words) within a line themselves are
sent right to left.

So, for the characters A, B, C, SPACE, D, E, F, SPACE, G, H, I
we might actually get the following appearance:  GHI DEF ABC

The stext extraction for this is then:

  <line>ABC</line>
  <line> </line>
  <line>DEF</line>
  <line> </line>
  <line>GHI</line>

which gives unexpected results when trying to select it.

As a workaround against this, we tweak the behaviour of the
stext device.

Typically as each character arrives, we consider adding it onto
the end of the "current line". If it doesn't fit, then we abandon
that line, and move to a new one.

The solution implemented here is to perform some extra processing.
Whenever we are finished with the current line, before we start a
new one, we check to see if the entirety of the current line fits
before one of the previous lines. If so, we join the two.

Clearly, we could go further with this processing and consider
whether the newly extended line then fits before/after other lines,
but this seems enough for now.

This gives sensible behaviour for text selection for this file,
and shouldn't hurt anything else. That's as much as we can hope
for, because cut-and-paste is clearly a lost cause thanks to
the broken ToUnicodes.