Fix problems when selecting/highlighting right to left text.
When selecting right to left text (such as trying to apply
a highlight to arabic text in mupdf-gl), the selection box
would behave very oddly.

If the selection point was in the right hand half of a glyph
we'd assume that it should be included, and in the left hand
half, that it should be excluded from the selection. This is
correct for l2r, but wrong for r2l text.

Fixed by considering the directionality of text.

The l2r/r2l direction detection could possibly be better as
it may trip over neutral stuff at the start of lines, but this
is MUCH better than before.

Also, when returning the list of highlight quads for a selection
we were assuming that successive chars would always 'merge'
new quads on the right. Again, this is wrong for r2l text.