Bug 706433: Move Page tree maps to be loaded/destroyed "just in time".
Recently, we'd moved the rev_page_map so that we dropped it
automatically when the file structure was modified. We
really ought to have done the same for the forward page map
too.

Also, the routine we were using to drop the map was using
a reference count which it assumed strictly paired map
loading with dropping. Remove that now.